pyproject.toml
[tool.isort]
profile = "black"
line_length = 79
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry]
name = "RH_ComfyUI"
version = "0.0.1"
description = "支持连接到ComfyUI实例或者连接至RunningHub代理的聊天机器人插件, 支持OneBot(QQ)、OneBotV12、QQ频道、微信、KOOK（开黑啦）、Telegram（电报）、FeiShu（飞书）、DoDo、Discord等多种Bot。"
authors = ["KimigaiiWuyi"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/KimigaiiWuyi/RH_ComfyUI"
repository = "https://github.com/KimigaiiWuyi/RH_ComfyUI"
documentation = "https://docs.sayu-bot.com/"
packages = [
    { include = "ZZZeroUID" }
]
exclude = ["tests", "deploy"]

[tool.poetry.dependencies]
python = "^3.10.11"

[[tool.poetry.source]]
name = "USTC"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = false
secondary = true

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^2.21.0"
pycln = "^2.1.2"

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"

[tool.black]
line-length = 79
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'
skip-string-normalization = true
extend-exclude = '''
'''

[tool.ruff]
line-length = 79
select = [
  "E", "W",   # pycodestyle
  "F",        # pyflakes
  # "I",        # isort
  "RUF",      # ruff
  "TRY",      # tryceratops
  "UP",

  # pylint
  "PLW",      # Warning
  "PLR",      # Refactor
  "PLE",      # Error

  "PTH",      # flake8-use-pathlib
  "SLF",      # flake8-self
  "RET",      # flake8-return
  "RSE",      # flake8-raise
  "T20",      # flake8-print
  "PIE",      # flake8-pie
  "ISC",      # flake8-implicit-str-concat
  "C4",       # flake8-comprehensions
  "COM",      # flake8-commas
  "A",        # flake8-builtins
  "B",        # flake8-bugbear
  "ASYNC",    # flake8-async
  "Q",        # flake8-quotes
]
ignore = [
  "ISC",
  "PLR2004",
  "A003",
  "E501",
  "COM812",
  "PLR0912",
  "PLR0915",
  "PLR0913",
  "PLR0911",
  "PLW0603",  # Using the global statement
  "TRY002",
  "TRY003"
]
exclude = [
  "gen.py",
  ".ruff_cache"
]
# Assume Python 3.8
target-version = "py310"

[tool.ruff.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.isort]
case-sensitive = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]
#force-wrap-aliases = true
combine-as-imports = true
order-by-type = true
relative-imports-order = "closest-to-furthest"
section-order = ["future", "standard-library", "first-party", "local-folder", "third-party"]

[tool.pdm]
[tool.pdm.build]
includes = []

[[tool.pdm.source]]
name = "USTC"
url = "https://mirrors.aliyun.com/pypi/simple/"
verify_ssl = true


[project]
name = "RH_ComfyUI"
version = "0.0.1"
description = "支持连接到ComfyUI实例或者连接至RunningHub代理的聊天机器人插件, 支持OneBot(QQ)、OneBotV12、QQ频道、微信、KOOK（开黑啦）、Telegram（电报）、FeiShu（飞书）、DoDo、Discord等多种Bot。"
authors = [
    {name = "KimigaiiWuyi", email = "444835641@qq.com"},
]
dependencies = [
    "beautifulsoup4>=4.12.2",
    "msgspec>=0.18.4",
    "httpx>=0.25.0",
    "pillow>=10.1.0",
    "aiofiles>=23.2.1",
    "aiohttp>=3.8.6",
    "qrcode[pil]>=7.4.2",
]
requires-python = ">=3.8.1,<4.0"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}


[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.276",
    "pre-commit>=3.3.2",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "pycln>=2.1.2"
]
test = [
    "nonebug>=0.3.0",
    "pytest>=7.2.0",
    "pytest-asyncio>=0.20.3"
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
